# Improve performance of EnumUtility generic methods