--- a/src/Hl7.Fhir.Base/Utility/EnumUtility.cs
+++ b/src/Hl7.Fhir.Base/Utility/EnumUtility.cs
@@ -27,9 +27,26 @@ namespace Hl7.Fhir.Utility
         public static string GetLiteral(this Enum e) => getEnumMapping(e.GetType()).GetLiteral(e);
 
         /// <summary>
+        /// Retrieves the literal value for the code represented by this enum <typeparamref name="T"/>, or the member name itself if there
+        /// is no literal value defined.
+        /// </summary>
+        public static string GetLiteral<T>(this T e) where T : struct, Enum => EnumMappingCache<T>.GetLiteral(e);
+
+        /// <summary>
+        /// Retrieves the literal value for the code represented by this nullable enum <typeparamref name="T"/>, or the member name itself if there
+        /// is no literal value defined, or null if the enum does not have a value.
+        /// </summary>
+        public static string? GetLiteral<T>(this T? e) where T : struct, Enum => e.HasValue ? e.Value.GetLiteral() : null;
+
+        /// <summary>
         /// Retrieves the system canonical for the code represented by this enum value, or <c>null</c> if there is no system defined.
         /// </summary>
         public static string? GetSystem(this Enum e) => e.GetAttributeOnEnum<EnumLiteralAttribute>()?.System ?? e.GetType().GetCustomAttribute<FhirEnumerationAttribute>()?.DefaultCodeSystem;
+
+        /// <summary>
+        /// Retrieves the system canonical for the code represented by this enum <typeparamref name="T"/>, or <c>null</c> if there is no system defined.
+        /// </summary>
+        public static string? GetSystem<T>(this T e) where T : struct, Enum => EnumMappingCache<T>.GetSystem(e);
 
         /// <summary>
         /// Retrieves the description for this enum value or the enumeration value itself if there is no description defined.
@@ -48,8 +65,8 @@ namespace Hl7.Fhir.Utility
         /// <summary>
         /// Finds an enumeration value from enum <typeparamref name="T"/> where the literal is the same as <paramref name="rawValue"/>.
         /// </summary>
-        public static T? ParseLiteral<T>(string? rawValue, bool ignoreCase = false) where T : struct
-            => (T?)(object?)ParseLiteral(rawValue, typeof(T), ignoreCase);
+        public static T? ParseLiteral<T>(string? rawValue, bool ignoreCase = false) where T : struct, Enum
+            => EnumMappingCache<T>.ParseLiteral(rawValue, ignoreCase);
 
         /// <summary>
         /// Gets the human readable name defined for the enumeration <paramref name="enumType"/>.
@@ -59,10 +76,67 @@ namespace Hl7.Fhir.Utility
         /// <summary>
         /// Gets the human readable name defined for the enumeration <typeparamref name="T"/>.
         /// </summary>
-        public static string GetName<T>() where T : struct => GetName(typeof(T));
+        public static string GetName<T>() where T : struct, Enum => EnumMappingCache<T>.Name;
 
         private static EnumMapping getEnumMapping(Type enumType)
             => CACHE.GetOrAdd(enumType, t => EnumMapping.Create(t));
+
+        private static class EnumMappingCache<TEnum>
+            where TEnum : struct, Enum
+        {
+            private static readonly Dictionary<string, TEnum> _literalToEnum = new();
+            private static readonly Dictionary<string, TEnum> _caseInsensitiveLiteralToEnum = new(StringComparer.OrdinalIgnoreCase);
+            private static readonly Dictionary<TEnum, string> _enumToLiteral = new();
+            private static readonly Dictionary<TEnum, string> _enumToSystem = new();
+
+            static EnumMappingCache()
+            {
+                var t = typeof(TEnum);
+                var enumAttr = t.GetTypeInfo().GetCustomAttribute<FhirEnumerationAttribute>();
+                Name = enumAttr?.BindingName ?? t.Name;
+                DefaultCodeSystem = enumAttr?.DefaultCodeSystem;
+                foreach (var enumValue in ReflectionHelper.FindEnumFields(t))
+                {
+                    var attr = ReflectionHelper.GetAttribute<EnumLiteralAttribute>(enumValue);
+                    string literal = attr?.Literal ?? enumValue.Name;
+
+                    var value = (TEnum)enumValue.GetValue(null)!;
+
+                    _enumToLiteral.Add(value, literal);
+                    _literalToEnum.Add(literal, value);
+                    _caseInsensitiveLiteralToEnum.Add(literal, value);
+                    if (attr?.System is string systemVal)
+                    {
+                        _enumToSystem.Add(value, systemVal);
+                    }
+                }
+            }
+
+            public static string Name { get; }
+
+            public static string? DefaultCodeSystem { get; }
+
+            public static string? GetSystem(TEnum value) =>
+                !_enumToSystem.TryGetValue(value, out string? result)
+                    ? DefaultCodeSystem
+                    : result;
+
+            public static string GetLiteral(TEnum value) =>
+                !_enumToLiteral.TryGetValue(value, out string? result)
+                    ? throw new InvalidOperationException($"Should only pass enum values that are member of the given enum: {value} is not a member of {Name}.")
+                    : result;
+
+            public static TEnum? ParseLiteral(string? literal, bool ignoreCase)
+            {
+                if (literal is null) return null;
+
+                var success = ignoreCase
+                    ? _caseInsensitiveLiteralToEnum.TryGetValue(literal, out TEnum result)
+                    : _literalToEnum.TryGetValue(literal, out result);
+
+                return success ? result : null;
+            }
+        }
 
         internal class EnumMapping
         {
@@ -79,7 +153,7 @@ namespace Hl7.Fhir.Utility
             public Type EnumType { get; private set; }
 
             private readonly Dictionary<string, Enum> _literalToEnum = new();
-            private readonly Dictionary<string, Enum> _lowercaseLiteralToEnum = new();
+            private readonly Dictionary<string, Enum> _caseInsensitiveLiteralToEnum = new(StringComparer.OrdinalIgnoreCase);
             private readonly Dictionary<Enum, string> _enumToLiteral = new();
 
             public string GetLiteral(Enum value) =>
@@ -92,7 +166,7 @@ namespace Hl7.Fhir.Utility
                 if (literal is null) return null;
 
                 var success = ignoreCase
-                    ? _lowercaseLiteralToEnum.TryGetValue(literal.ToLowerInvariant(), out Enum? result)
+                    ? _caseInsensitiveLiteralToEnum.TryGetValue(literal, out Enum? result)
                     : _literalToEnum.TryGetValue(literal, out result);
 
                 return success ? result : null;
@@ -113,7 +187,7 @@ namespace Hl7.Fhir.Utility
 
                     result._enumToLiteral.Add(value, literal);
                     result._literalToEnum.Add(literal, value);
-                    result._lowercaseLiteralToEnum.Add(literal.ToLowerInvariant(), value);
+                    result._caseInsensitiveLiteralToEnum.Add(literal, value);
                 }
 
                 return result;