--- a/src/Hl7.Fhir.Support.Tests/Utility/EnumMappingTest.cs
+++ b/src/Hl7.Fhir.Support.Tests/Utility/EnumMappingTest.cs
@@ -9,6 +9,7 @@
 using FluentAssertions;
 using Hl7.Fhir.Utility;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
 using System.Diagnostics;
 
 namespace Hl7.Fhir.Support.Tests.Utils
@@ -33,6 +34,25 @@ namespace Hl7.Fhir.Support.Tests.Utils
             Assert.AreEqual("ItemThree", TestEnum.Item3.GetLiteral());
             Assert.AreEqual("Item3", TestEnum.Item3.GetDocumentation());
             Assert.IsNull(TestEnum.Item3.GetSystem());
+        }
+
+        [TestMethod]
+        public void TestCreationNonGeneric()
+        {
+            Assert.AreEqual("Testee", EnumUtility.GetName(typeof(TestEnum)));
+            Assert.AreEqual(TestEnum.Item1, EnumUtility.ParseLiteral("Item1", typeof(TestEnum)));
+            Assert.IsNull(EnumUtility.ParseLiteral("Item2", typeof(TestEnum)));
+            Assert.AreEqual(TestEnum.Item2, EnumUtility.ParseLiteral("ItemTwo", typeof(TestEnum)));
+            Assert.IsNull(EnumUtility.ParseLiteral("iTeM1", typeof(TestEnum)));
+            Assert.AreEqual(TestEnum.Item1, EnumUtility.ParseLiteral("iTeM1", typeof(TestEnum), ignoreCase: true));
+
+            Assert.AreEqual("Item1", ((Enum)TestEnum.Item1).GetLiteral());
+            Assert.AreEqual("ItemTwo", ((Enum)TestEnum.Item2).GetLiteral());
+            Assert.AreEqual("This is item two", TestEnum.Item2.GetDocumentation());
+            Assert.AreEqual("http://example.org/test-system", ((Enum)TestEnum.Item2).GetSystem());
+            Assert.AreEqual("ItemThree", ((Enum)TestEnum.Item3).GetLiteral());
+            Assert.AreEqual("Item3", ((Enum)TestEnum.Item3).GetDocumentation());
+            Assert.IsNull(((Enum)TestEnum.Item3).GetSystem());
         }
 
         [FhirEnumeration("Testee")]
@@ -108,6 +128,21 @@ namespace Hl7.Fhir.Support.Tests.Utils
         }
 
         [TestMethod]
+        public void EnumParsingPerformanceNonGeneric()
+        {
+            var sw = new Stopwatch();
+            sw.Start();
+
+            for (var i = 0; i < 10000; i++)
+                EnumUtility.ParseLiteral("male", typeof(TestAdministrativeGender));
+
+            sw.Stop();
+
+            Debug.WriteLine(sw.ElapsedMilliseconds);
+            Assert.IsTrue(sw.ElapsedMilliseconds < 100);
+        }
+
+        [TestMethod]
         public void TestEnumMapping()
         {
             Assert.AreEqual(TestAdministrativeGender.Male, EnumUtility.ParseLiteral<TestAdministrativeGender>("male"));
@@ -119,13 +154,41 @@ namespace Hl7.Fhir.Support.Tests.Utils
         }
 
         [TestMethod]
+        public void TestEnumMappingNonGeneric()
+        {
+            Assert.AreEqual(TestAdministrativeGender.Male, EnumUtility.ParseLiteral("male", typeof(TestAdministrativeGender)));
+            Assert.IsNull(EnumUtility.ParseLiteral("maleX", typeof(TestAdministrativeGender)));
+            Assert.AreEqual(X.a, EnumUtility.ParseLiteral("a", typeof(X)));
+
+            Assert.AreEqual("Male", TestAdministrativeGender.Male.GetDocumentation());
+            Assert.AreEqual("a", X.a.GetDocumentation()); // default documentation = name of item
+        }
+
+        [TestMethod]
         public void EnumLiteralPerformance()
         {
             var result = string.Empty;
+            var value = TestAdministrativeGender.Male;
 
             var sw = Stopwatch.StartNew();
             for (var i = 0; i < 50_000; i++)
-                result = TestAdministrativeGender.Male.GetLiteral();
+                result = value.GetLiteral();
+            sw.Stop();
+
+            Assert.AreEqual("male", result);
+            Debug.WriteLine(sw.ElapsedMilliseconds);
+            Assert.IsTrue(sw.ElapsedMilliseconds < 500);
+        }
+
+        [TestMethod]
+        public void EnumLiteralPerformanceNonGeneric()
+        {
+            var result = string.Empty;
+            Enum value = TestAdministrativeGender.Male;
+
+            var sw = Stopwatch.StartNew();
+            for (var i = 0; i < 50_000; i++)
+                result = value.GetLiteral();
             sw.Stop();
 
             Assert.AreEqual("male", result);