## Description
- Improve performance of generic methods in EnumUtility
- Remove usage of ToLowerInvariant for case insensitive parsing in EnumUtility
- Added generic version of GetLiteral and GetSystem methods in EnumUtility
- Added EnumUtility benchmarks
- `ParseLiteral<T>` and `GetName<T>` have an added Enum generic type constraint which is technically a breaking change, but these methods are unlikely to be called with non-Enum generic type parameters.

## Related issues
Resolves #2575.

## Testing
- Hl7.Fhir.Support.Tests.Utils.EnumMappingTest
- Firely.Sdk.Benchmarks.EnumUtilityBenchmarks

```
BenchmarkDotNet v0.13.8, Windows 11 (10.0.22621.2283/22H2/2022Update/SunValley2)
AMD Ryzen 7 7840U w/ Radeon 780M Graphics, 1 CPU, 16 logical and 8 physical cores
.NET SDK 7.0.401
  [Host]     : .NET 6.0.22 (6.0.2223.42425), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.22 (6.0.2223.42425), X64 RyuJIT AVX2
```
| Method                                      | Mean         | Error      | StdDev     | Gen0   | Allocated |
|-------------------------------------------- |-------------:|-----------:|-----------:|-------:|----------:|
| EnumUtilityGetLiteral                       |     7.557 ns |  0.0530 ns |  0.0470 ns |      - |         - |
| EnumUtilityGetLiteralNonGeneric             |    46.648 ns |  0.4639 ns |  0.4340 ns |      - |         - |
| EnumUtilityParseLiteral                     |    19.346 ns |  0.1596 ns |  0.1493 ns |      - |         - |
| EnumUtilityParseLiteralNonGeneric           |    29.049 ns |  0.1169 ns |  0.0976 ns |      - |         - |
| EnumUtilityParseLiteralIgnoreCase           |    21.979 ns |  0.2612 ns |  0.2316 ns |      - |         - |
| EnumUtilityParseLiteralIgnoreCaseNonGeneric |    32.509 ns |  0.1329 ns |  0.1244 ns |      - |         - |
| EnumUtilityGetSystem                        |     2.728 ns |  0.1049 ns |  0.0981 ns |      - |         - |
| EnumUtilityGetSystemNonGeneric              | 2,882.727 ns | 28.1981 ns | 24.9969 ns | 0.1030 |     864 B |

## FirelyTeam Checklist
- [ ] **Update the title** of the PR to be succinct and less than 50 characters
- [ ] Mark the PR with the label **breaking change** when this PR introduces breaking changes