@mmsmits 2023-09-26
Hi @almostchristian , Thanks for the PR!
The build didn't succeed earlier because of a pipeline error at our end. That's fixed now.
I've pulled your PR, this is super helpful.